Introduce a back-end server
