+++ src/Backend/App.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const app = express();
+const port = 3000;
+
+app.get("/", (req, res) => res.send("Hello World!"));
+
+app.listen(port, () => console.log(`Example app listening on port ${port}!`));
