--- src/Details/Details.css
@@ -1,5 +1,0 @@
-.container .text {
-  margin: 5px 20px 0 30px;
-  letter-spacing: 1px;
-  line-height: 1.6;
-}
